fix: resolve all 13 Phase 1A QA issues
Critical:
- C-1: Replace placeholder badge images with real partner badges (Microsoft, Google, QuickBooks)
- C-2: Fix comparison table visibility — explicit colors replace theme tokens for CSS contrast
- C-3: Replace hero right-panel placeholder with stats card (services, turnaround, accuracy, location)

Medium:
- M-1: Fix hero pricing — use flagship product price instead of lowest across catalog
- M-2: Fix FAQ section rendering — replace Card with explicit div+colors on subscription page
- M-3: Add Waitlist page link to QB portal footer navigation
- M-4: Fix IT Consulting page title mismatch (Managed IT → Project Management)
- M-5: Reduce whitespace gap between service cards and Why Choose section (!pt-8)

Minor:
- m-1: Unify blog date formatting (month: long → short for featured post)
- m-2: Add 'Contact for Pricing' to QuickBooks services without listed prices
- m-3: Remove duplicate CTA on IT Consulting page bottom
- m-4: Change '24h' stat to '< 24h Response SLA' for clarity
- m-5: Fix hero card right-edge viewport clipping (max-w-[40%])

+++ artifacts/nexfortis/src/pages/blog.tsx
@@ -171,7 +171,7 @@ export default function Blog() {
                     <span className="flex items-center gap-1.5">
                       <Calendar className="w-3.5 h-3.5" aria-hidden="true" />
                       <time dateTime={featured.createdAt}>
-                        {new Date(featured.createdAt).toLocaleDateString("en-CA", { year: "numeric", month: "long", day: "numeric" })}
+                        {new Date(featured.createdAt).toLocaleDateString("en-CA", { year: "numeric", month: "short", day: "numeric" })}
                       </time>
                     </span>
                     <span className="flex items-center gap-1.5">
